Refactor GlideApp package
